Refactoring Node
